requirements.txt
langchain
langchain-ollama
langchain-community
langgraph
faiss-cpu
python-dotenv